requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
django-phonenumber-field==6.1.0
djangorestframework==3.13.1
phonenumbers==8.12.33
djangorestframework-simplejwt==5.2.0
django-filter==22.1
markdown==3.3.7